+++ examples/mobc/src/src_user/component_driver/com/gs.c
@@ -26,7 +26,7 @@
 
 // それぞれ AD, BD, BC
 static uint8_t GS_rx_header_[GS_RX_HEADER_NUM][GS_RX_HEADER_SIZE];
-static uint8_t GS_tx_frame_[VCDU_LEN];
+static uint8_t GS_tx_frame_[AOSTF_LEN];
 
 /**
  * @brief CCSDS 側 Driver の DS 上での初期化設定
@@ -114,8 +114,8 @@ CDS_INIT_ERR_CODE GS_init(GS_Driver* gs_driver,
     gs_driver->info[i].rx.tctf.bd_rec_status = GS_VALIDATE_ERR_OK;
 
     gs_driver->info[i].tx.send_cycle = 0;
-    gs_driver->info[i].tx.vcid = VCDU_VCID_UNKNOWN;
-    gs_driver->info[i].tx.vcdu_counter = 0;
+    gs_driver->info[i].tx.vcid = AOSTF_VCID_UNKNOWN;
+    gs_driver->info[i].tx.aostf_counter = 0;
   }
 
   gs_driver->ccsds_info.buffer_num = 8;
@@ -152,7 +152,7 @@ static void GS_load_default_driver_super_init_settings_(ComponentDriverSuper* p_
     CDSSC_enable_strict_frame_search(p_stream_config);   // 複数ストリームがあり，かつ論理的な受信漏れを0にするため．
 
     CDSSC_set_tx_frame(p_stream_config, GS_tx_frame_); // 全 stream, uart, ccsds に関わらず共通
-    CDSSC_set_tx_frame_size(p_stream_config, VCDU_LEN); // VCDU を送信
+    CDSSC_set_tx_frame_size(p_stream_config, AOSTF_LEN); // AOSTF を送信
 
     CDSSC_set_rx_header(p_stream_config, GS_rx_header_[stream], GS_RX_HEADER_SIZE);
     CDSSC_set_rx_frame_size(p_stream_config, -1); // 可変長
@@ -254,21 +254,21 @@ static CDS_ERR_CODE GS_analyze_rec_data_(CDS_StreamConfig* p_stream_config, void
   return CDS_ERR_CODE_OK;
 }
 
-CDS_CMD_ERR_CODE GS_send_vcdu(GS_Driver* gs_driver, const VCDU* vcdu)
+CDS_CMD_ERR_CODE GS_send_aostf(GS_Driver* gs_driver, const AosTransferFrame* aostf)
 {
   CDS_ERR_CODE ret_ccsds = CDS_ERR_CODE_OK;
   CDS_ERR_CODE ret_uart  = CDS_ERR_CODE_OK;
-  size_t vcdu_size = sizeof(VCDU);
+  size_t aostf_size = sizeof(AosTransferFrame);
 
   // パディングが無ければ元を GS_tx_frame_ にコピーさせる (444Byte) のコピーが無駄
-  if (vcdu_size == VCDU_LEN)
+  if (aostf_size == AOSTF_LEN)
   {
-    CDSSC_set_tx_frame(&gs_driver->driver_ccsds.super.stream_config[GS_TX_STREAM], (uint8_t*)vcdu);
-    CDSSC_set_tx_frame(&gs_driver->driver_uart.super.stream_config[GS_TX_STREAM], (uint8_t*)vcdu);
+    CDSSC_set_tx_frame(&gs_driver->driver_ccsds.super.stream_config[GS_TX_STREAM], (uint8_t*)aostf);
+    CDSSC_set_tx_frame(&gs_driver->driver_uart.super.stream_config[GS_TX_STREAM], (uint8_t*)aostf);
   }
   else
   {
-    VCDU_generate_byte_stream(vcdu, GS_tx_frame_); // 送信元にセット 消したいなぁ...
+    AOSTF_generate_byte_stream(aostf, GS_tx_frame_); // 送信元にセット 消したいなぁ...
     CDSSC_set_tx_frame(&gs_driver->driver_ccsds.super.stream_config[GS_TX_STREAM], GS_tx_frame_);
     CDSSC_set_tx_frame(&gs_driver->driver_uart.super.stream_config[GS_TX_STREAM], GS_tx_frame_);
   }
@@ -278,8 +278,8 @@ CDS_CMD_ERR_CODE GS_send_vcdu(GS_Driver* gs_driver, const VCDU* vcdu)
   if (gs_driver->ccsds_info.buffer_num)
   {
     gs_driver->info[GS_PORT_TYPE_CCSDS].tx.send_cycle = TMGR_get_master_total_cycle();
-    gs_driver->info[GS_PORT_TYPE_CCSDS].tx.vcid = VCDU_get_vcid(vcdu);
-    gs_driver->info[GS_PORT_TYPE_CCSDS].tx.vcdu_counter = VCDU_get_vcdu_counter(vcdu);
+    gs_driver->info[GS_PORT_TYPE_CCSDS].tx.vcid = AOSTF_get_vcid(aostf);
+    gs_driver->info[GS_PORT_TYPE_CCSDS].tx.aostf_counter = AOSTF_get_aostf_counter(aostf);
 
     // DS 側の名称が cmd なだけで送信しているのは TLM
     ret_ccsds = CDS_send_general_cmd(&gs_driver->driver_ccsds.super, GS_TX_STREAM);
@@ -289,8 +289,8 @@ CDS_CMD_ERR_CODE GS_send_vcdu(GS_Driver* gs_driver, const VCDU* vcdu)
   if (gs_driver->driver_uart.is_tlm_on)
   {
     gs_driver->info[GS_PORT_TYPE_UART].tx.send_cycle = TMGR_get_master_total_cycle();
-    gs_driver->info[GS_PORT_TYPE_UART].tx.vcid = VCDU_get_vcid(vcdu);
-    gs_driver->info[GS_PORT_TYPE_UART].tx.vcdu_counter = VCDU_get_vcdu_counter(vcdu);
+    gs_driver->info[GS_PORT_TYPE_UART].tx.vcid = AOSTF_get_vcid(aostf);
+    gs_driver->info[GS_PORT_TYPE_UART].tx.aostf_counter = AOSTF_get_aostf_counter(aostf);
 
     // DS 側の名称が cmd なだけで送信しているのは TLM
     ret_uart  = CDS_send_general_cmd(&gs_driver->driver_uart.super,  GS_TX_STREAM);